Complete hashmaps in rust. Same concept as other language. Note that I remember is that it if you put in var: String as keys then the dictionary took ownership. That is a bit odd. Other than that they behaved the same as I thought. I love hashmaps.
Then i did the quiz. It was confusing at first because I did not
understand the aim. But it stated vec<String> which I did not
understand how it could be compared to a list of string slices.
That still does not make super good sense. But I guess it is still
just bytes. The string slices are on the stack and pre-determined,
while the String is on the heap.

Either way I use match and a lot of documentation look up. Did not
use chatgpt for anything and I generally progress a lot faster then.

Either way that was today. Next day it is time for Options.
There seems to be some tools with  if-let. Error handling essentially.
Which I honestly think is the next step for me as developer inside
and outside of Rust. Meaning C aswell. I need to understand the
necessity to truly be good.