Man... I got burned out by return value should be [1,2,3,4] because it made me think it was supposed to be an array instead it could be an Vec instead. Which messed with my head. Also, collect() is more powerful than I thought in gather the data into the correct data structure. damn. So take away is to trust it. Also gpt help was not helping. It uses unwrap() for every- thing. Which is crazy. Like we need match and clone and all kinds of weird things. It was not productive to use it here at all.
When it comes to traits I get them now mostly. Just need
practice. But the current challenge is understanding
when to use a reference and when to use a value. But
the lifetime concept generally comes alright to me.
tests are neat but again just gotta write them which is
the challenege for all developers around the world.