Threads. Nice. Love them and this is what I want to get into I want to understand the depth of them and how to utilize them alongside with how to utilize cores best. There is a lot of functionality like how to manage properly and how to use properly. But I have concepts down. And whatever thing I see I usually pick it up.
but now rust specific. Only thing again is to make sure you satisfy
the rust borrow. It is the thing that wil make you confused.
But closures are actually the thing that confused me more.
But it makes sense a bit now.
But I wonder if I can make the thread run a sequence of functions call
Tbh, there seems to be nothing to stop it but the examples
have been quite simple so far. In book and here.
But that will be test worthy. Either way. Good job mate :)